Move and simplify tests
