+++ framework/test/jest/unit/specs/modules/chain/bft/bft.spec.js
@@ -14,11 +14,6 @@
 
 'use strict';
 
-const { Mnemonic } = require('@liskhq/lisk-passphrase');
-const {
-	getAddressAndPublicKeyFromPassphrase,
-} = require('@liskhq/lisk-cryptography');
-
 const {
 	Block: blockFixture,
 } = require('../../../../../../mocha/fixtures/blocks');
@@ -388,126 +383,6 @@ describe('bft', () => {
 					403 - activeDelegates * 2,
 				);
 				expect(storageMock.entities.Block.get).toHaveBeenCalledTimes(0);
-			});
-		});
-
-		describe('computeHeadersForNewBlock()', () => {
-			let bft;
-
-			beforeEach(async () => {
-				bft = new BFT(bftParams);
-
-				storageMock.entities.Block.get.mockReset();
-				storageMock.entities.Block.get.mockReturnValue([]);
-
-				await bft.init();
-			});
-
-			it('should maxHeightPreviouslyForged and prevotedConfirmedUptoHeight properties in an object literal', async () => {
-				const headers = await bft.computeHeadersForNewBlock('aDelegatePubKey');
-
-				expect(headers).toHaveProperty('maxHeightPreviouslyForged');
-				expect(headers).toHaveProperty('prevotedConfirmedUptoHeight');
-				expect(headers).toEqual({
-					maxHeightPreviouslyForged: 0,
-					prevotedConfirmedUptoHeight: 0,
-				});
-			});
-
-			it('should compute a correct value for maxHeightPreviouslyForged, returning the height of the last block a delegate with the given public key forged a block', async () => {
-				// Arrange
-				const {
-					publicKey: forgingDelegatePubKey,
-				} = getAddressAndPublicKeyFromPassphrase(Mnemonic.generateMnemonic());
-
-				const blockHeaders = [
-					{
-						blockId: '1234567',
-						height: 1,
-						maxHeightPreviouslyForged: 0,
-						prevotedConfirmedUptoHeight: 0,
-						activeSinceRound: 0,
-						delegatePublicKey: forgingDelegatePubKey,
-					},
-					{
-						blockId: '1234568',
-						height: 2,
-						maxHeightPreviouslyForged: 0,
-						prevotedConfirmedUptoHeight: 0,
-						activeSinceRound: 0,
-						delegatePublicKey: getAddressAndPublicKeyFromPassphrase(
-							Mnemonic.generateMnemonic(),
-						).publicKey,
-					},
-					{
-						blockId: '1234569',
-						height: 3,
-						maxHeightPreviouslyForged: 1,
-						prevotedConfirmedUptoHeight: 0,
-						activeSinceRound: 0,
-						delegatePublicKey: forgingDelegatePubKey,
-					},
-				];
-
-				blockHeaders.forEach(blockHeader => {
-					bft.finalityManager.addBlockHeader(blockHeader);
-				});
-
-				// Act
-				const headers = await bft.computeHeadersForNewBlock(
-					forgingDelegatePubKey,
-				);
-
-				// Assert
-				expect(headers).toEqual({
-					maxHeightPreviouslyForged: 3,
-					prevotedConfirmedUptoHeight: 0,
-				});
-			});
-
-			it("should compute set maxHeightPreviouslyForged to 0 if it's the first time the delegate is forging a block", async () => {
-				// Arrange
-				const {
-					publicKey: forgingDelegatePubKey,
-				} = getAddressAndPublicKeyFromPassphrase(Mnemonic.generateMnemonic());
-
-				const blockHeaders = [
-					{
-						blockId: '1234568',
-						height: 1,
-						maxHeightPreviouslyForged: 0,
-						prevotedConfirmedUptoHeight: 0,
-						activeSinceRound: 0,
-						delegatePublicKey: getAddressAndPublicKeyFromPassphrase(
-							Mnemonic.generateMnemonic(),
-						).publicKey,
-					},
-					{
-						blockId: '1234569',
-						height: 2,
-						maxHeightPreviouslyForged: 0,
-						prevotedConfirmedUptoHeight: 0,
-						activeSinceRound: 0,
-						delegatePublicKey: getAddressAndPublicKeyFromPassphrase(
-							Mnemonic.generateMnemonic(),
-						).publicKey,
-					},
-				];
-
-				blockHeaders.forEach(blockHeader => {
-					bft.finalityManager.addBlockHeader(blockHeader);
-				});
-
-				// Act
-				const headers = await bft.computeHeadersForNewBlock(
-					forgingDelegatePubKey,
-				);
-
-				// Assert
-				expect(headers).toEqual({
-					maxHeightPreviouslyForged: 0,
-					prevotedConfirmedUptoHeight: 0,
-				});
 			});
 		});
